requirements.txt
Django==5.1.6
djangorestframework==3.15.2
pytz==2023.3.post1
